pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biblioteca"
version = "0.0.2"
authors = [
  { name="Thiago Pacheco Rocha", email="thiago.rocha@ufu.br" },
]
description = "Sistema de biblioteca para projeto da disciplina sistemas distribuídos."
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
   "grpcio-tools",
   "pysyncobj",
   "multimethod",
   "jsonpickle",
]

[project.optional-dependencies]
dev = []

[project.scripts]
cad-server="biblioteca.cad.cad_server:run"
cad-client="biblioteca.cad.cad_client:run"
bib-server="biblioteca.bib.bib_server:run"
bib-client="biblioteca.bib.bib_client:run"
db-server="biblioteca.leveldb.db_server:run"

[project.urls]
Homepage = "https://github.com/costadev00/sistemas-distribuidos-projeto"
Issues = "https://github.com/costadev00/sistemas-distribuidos-projeto/issues"
